Add tracing commong to cja and re-export it for other apps to use
