+++ tracing-common/Cargo.toml
@@ -6,6 +6,15 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+sentry = { version = "0.31.5", default-features = false, features = [
+  "rustls",
+  "backtrace",
+  "contexts",
+  "panic",
+  "tower",
+  "reqwest",
+] }
+sentry-tracing = "0.31.8"
 tracing = { workspace = true }
 tracing-opentelemetry = { workspace = true }
 tracing-subscriber = { workspace = true }
